Modified counting, radix sort. Added new test case
Counting sort before could support up to 50 numbers because of the nega-
tive offset. I originally intended it to support 100, so I changed the
size of the counting array to 200, so it can support numbers from -100
to 100. Radix sort now splits the array into negative and positive numb-
ers, does radix sort on both side, and appends them together before sub-
mitting. While the implementation on this is not perfect and is a bit
hardcoded, it is helpful to practice these again